options: handle alias options differently
When setting an alias option, its value is not saved in the file-local
options case. The ensure_backup function in m_config.c exits if it
encounters an aliased option, because it should not be saved
additionally to the original option. However, the original option is
likely never saved in this case.

Change it so that ensure_backup always accesses the original option.
The original option is the one that first appears in the option list.