rename osd_render -> draw_bmp
The naming was conflicting with osd_libass.c and osd_dummy.c. Also,
there's an osd_render member in struct osd_state.

Rename it to draw_bmp, which also reflects the fact a bit better that
it's on the end of the render chain, and that it's not actually where
OSD text is rendered to sub-bitmaps.

Not using draw_osd, because the exported function doesn't draw the
whole OSD, only a part of it.